requirements.txt
git+https://github.com/pytorch/tnt.git@master
fire
ipdb
torchvision
visdom
tqdm
